foreign key, modified posts for owner_id

+++ app/routers/post.py
@@ -21,9 +21,8 @@ async def create_post(post: schemas.PostCreate,
                       db: Session = Depends(get_db), 
                       current_user: int = Depends(oauth2.get_current_user)):
     
-    print(current_user.email)
-    post.dict()
-    new_post = models.Post(**post.dict())
+
+    new_post = models.Post(owner_id=current_user.id, **post.dict())
 
     db.add(new_post)
     db.commit()
@@ -47,10 +46,14 @@ def delete_post(id: int, db: Session = Depends(get_db),
                 current_user: int = Depends(oauth2.get_current_user)):
     
     post_query = db.query(models.Post).filter(models.Post.id==id)
-    
+    post = post_query.first()
     if post_query.first() == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
                             detail=f'post with id: {id} does not exist')
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, 
+                            detail=f'Not authorized to perform requested actions')
+
     post_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
@@ -66,6 +69,9 @@ def update_post(id: int, updated_post: schemas.PostCreate,
     if post == None: 
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
                             detail=f'post with id: {id} does not exist')
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, 
+                            detail=f'Not authorized to perform requested actions')
     
     post_query.update(updated_post.dict(), synchronize_session=False)
     db.commit()
